Check for empty ldr/p2 lib ids (these happened)
https://github.com/douglasnaphas/madliberation/issues/200

+++ itest.sh
@@ -73,8 +73,18 @@ echo "P2_NAME: ${P2_NAME}"
 echo ${P2_NAME}
 LDR_LIB_ID=$(echo "${ITEMS}" | \
   jq '.[] | select((.["game_name"]["S"] == "'"${LDR_NAME}"'") and (.["lib_id"]["S"] | startswith("participant#"))) | .["lib_id"]["S"]')
+if [[ -z ${LDR_LIB_ID} ]]
+then
+  echo "empty LDR_LIB_ID: ${LDR_LIB_ID}"
+  exit 1
+fi
 P2_LIB_ID="$(echo "${ITEMS}" | \
   jq '.[] | select((.["game_name"]["S"] == "'"${P2_NAME}"'") and (.["lib_id"]["S"] | startswith("participant#"))) | .["lib_id"]["S"]')"
+if [[ -z ${P2_LIB_ID} ]]
+then
+  echo "empty P2_LIB_ID: ${P2_LIB_ID}"
+  exit 1
+fi
 # delete leader
 aws dynamodb delete-item \
   --table-name seders \
